Updated battle estimates to better estimate damage of weapons when moving
