+++ server/battles.go
@@ -82,6 +82,17 @@ func (s *server) testBattle(w http.ResponseWriter, r *http.Request) {
 				{HullComponent: cs.BetaTorpedo.Name, HullSlotIndex: 4, Quantity: 1},
 				{HullComponent: cs.BetaTorpedo.Name, HullSlotIndex: 3, Quantity: 1},
 			}),
+		cs.NewShipDesign(player1, 7).
+			WithName("INT-1MF-SCP - Dragonfly").
+			WithHull(cs.Destroyer.Name).
+			WithSlots([]cs.ShipDesignSlot{
+				{HullComponent: cs.FuelMizer.Name, HullSlotIndex: 1, Quantity: 1},
+				{HullComponent: cs.FuelTank.Name, HullSlotIndex: 6, Quantity: 1},
+				{HullComponent: cs.BattleComputer.Name, HullSlotIndex: 7, Quantity: 1},
+				{HullComponent: cs.PulsedSapper.Name, HullSlotIndex: 2, Quantity: 1},
+				{HullComponent: cs.BattleComputer.Name, HullSlotIndex: 4, Quantity: 1},
+				{HullComponent: cs.PhaserBazooka.Name, HullSlotIndex: 3, Quantity: 1},
+			}),
 	)
 
 	player2.Designs = append(player2.Designs,
@@ -142,6 +153,10 @@ func (s *server) testBattle(w http.ResponseWriter, r *http.Request) {
 				{HullComponent: cs.ColloidalPhaser.Name, HullSlotIndex: 3, Quantity: 1},
 				{HullComponent: cs.BearNeutrinoBarrier.Name, HullSlotIndex: 4, Quantity: 1},
 			}),
+		cs.NewShipDesign(player2, 7).
+			WithName("Empty Spacedock").
+			WithHull(cs.SpaceDock.Name).
+			WithSlots([]cs.ShipDesignSlot{}),
 	)
 
 	p1BattleCruiser := player1.Designs[0]
@@ -149,21 +164,25 @@ func (s *server) testBattle(w http.ResponseWriter, r *http.Request) {
 	p1StalwartDefender := player1.Designs[2]
 	p1AcceleratorPlatform := player1.Designs[3]
 	p1DestroyerV1 := player1.Designs[5]
+	p1DragonFly := player1.Designs[6]
 	p2Teamster := player2.Designs[0]
 	p2Scout := player2.Designs[1]
 	p2StalwartDefender := player2.Designs[2]
 	p2Sapper := player2.Designs[3]
 	p2Frigate := player2.Designs[4]
 	p2DDS10 := player2.Designs[5]
+	p2EmptySpaceDock := player2.Designs[6]
 	_ = p1BattleCruiser
 	_ = p1Scout
 	_ = p1StalwartDefender
 	_ = p1AcceleratorPlatform
+	_ = p1DestroyerV1
 	_ = p2Teamster
 	_ = p2Scout
 	_ = p2StalwartDefender
 	_ = p2Sapper
 	_ = p2Frigate
+	_ = p2DDS10
 
 	fleets := []*cs.Fleet{
 		// player1's tokens
@@ -174,8 +193,8 @@ func (s *server) testBattle(w http.ResponseWriter, r *http.Request) {
 			BaseName: "Fleet1",
 			Tokens: []cs.ShipToken{
 				{
-					DesignNum: p1DestroyerV1.Num,
-					Quantity:  5,
+					DesignNum: p1DragonFly.Num,
+					Quantity:  1,
 				},
 			},
 		},
@@ -192,7 +211,7 @@ func (s *server) testBattle(w http.ResponseWriter, r *http.Request) {
 				},
 				{
 					Quantity:  1,
-					DesignNum: p2DDS10.Num,
+					DesignNum: p2EmptySpaceDock.Num,
 				},
 			},
 		}}
